feat: 리뷰사항 반영. 모달분리 & 디테일페이지 제목제거
